new file for headerstripes, export files, fixed
